Refactor: Consolidate macro imports into `macros` module and update usage across the codebase
