+++ .changeset/hide_internal_macros_from_exported_code.md
@@ -1,0 +1,5 @@
+---
+default: patch
+---
+
+# Hide internal macros from exported code.
