+++ src/f64/math.rs
@@ -1,5 +1,5 @@
+use crate::macros::math;
 use super::{GuardedF64, UnguardedF64};
-use crate::math;
 
 math!(
     (GuardedF64, UnguardedF64)
