+++ src/f64/ops_binary.rs
@@ -1,21 +1,21 @@
 use super::{GuardedF64, UnguardedF64};
-use crate::binary_operation;
+use crate::macros::binary_operation;
 use std::ops::{Add, Div, Mul, Rem, Sub};
 
 binary_operation!(
     impl Add for ...(GuardedF64, UnguardedF64) {
         r"
             Adds two `GuardedF64` values or a `GuardedF64` and a `f64`.
-    
+
             # Example
-    
+
             ```rust
             use floatguard::{GuardedF64, FloatError};
-    
+
             let value1 = GuardedF64::new(2.0).unwrap();
             let value2 = GuardedF64::new(3.0).unwrap();
             assert_eq!((value1 + value2).check().unwrap(), 5.0);
-    
+
             assert_eq!((value1 + f64::NAN).check(), Err(FloatError::NaN));
             ```
         "
@@ -29,16 +29,16 @@ binary_operation!(
     impl Sub for ...(GuardedF64, UnguardedF64) {
         r"
             Subtracts one `GuardedF64` value from another or a `f64` from a `GuardedF64`.
-    
+
             # Example
-    
+
             ```rust
             use floatguard::{GuardedF64, FloatError};
-    
+
             let value1 = GuardedF64::new(5.0).unwrap();
             let value2 = GuardedF64::new(3.0).unwrap();
             assert_eq!(f64::try_from(value1 - value2), Ok(2.0));
-    
+
             assert_eq!((value1 - f64::NAN).check(), Err(FloatError::NaN));
             ```
         "
@@ -52,16 +52,16 @@ binary_operation!(
     impl Mul for ...(GuardedF64, UnguardedF64) {
         r"
             Multiplies two `GuardedF64` values or a `GuardedF64` and a `f64`.
-    
+
             # Example
-    
+
             ```rust
             use floatguard::{GuardedF64, FloatError};
-    
+
             let value1 = GuardedF64::new(2.0).unwrap();
             let value2 = GuardedF64::new(3.0).unwrap();
             assert_eq!(f64::try_from(value1 * value2), Ok(6.0));
-    
+
             assert_eq!((value1 * f64::NAN).check(), Err(FloatError::NaN));
             ```
         "
@@ -75,27 +75,27 @@ binary_operation!(
     impl Div for ...(GuardedF64, UnguardedF64) {
         r"
             Divides one `GuardedF64` value by another or a `f64` by a `GuardedF64`.
-    
+
             # Example
-    
+
             ```rust
             use floatguard::{GuardedF64, UnguardedF64, FloatError};
-    
+
             let value1 = GuardedF64::new(6.0).unwrap();
             let value2 = GuardedF64::new(3.0).unwrap();
             assert_eq!(f64::try_from(value1 / value2), Ok(2.0));
-    
+
             let value1 = UnguardedF64::new(6.0);
             assert_eq!((value1 / 0.0).check(), Err(FloatError::Infinity));
-    
+
             let value1 = UnguardedF64::new(6.0);
             assert_eq!((value1 / f64::NAN).check(), Err(FloatError::NaN));
             assert_eq!((f64::NAN / value1).check(), Err(FloatError::NaN));
-    
+
             let value1 = UnguardedF64::new(6.0);
             assert_eq!((f64::INFINITY / value1).check(), Err(FloatError::Infinity));
             assert_eq!((value1 / f64::INFINITY).check(), Err(FloatError::Infinity));
-    
+
             let value1 = UnguardedF64::new(f64::INFINITY);
             let value2 = UnguardedF64::new(f64::NAN);
             assert_eq!((value1 / value2).check(), Err(FloatError::NaN));
@@ -121,29 +121,29 @@ binary_operation!(
         r"
             Computes the remainder of division between two `GuardedF64` values or a `GuardedF64` and
             a `f64`.
-    
+
             # Example
-    
+
             ```rust
             use floatguard::{GuardedF64, UnguardedF64, FloatError};
-    
+
             let value1 = GuardedF64::new(5.0).unwrap();
             let value2 = GuardedF64::new(3.0).unwrap();
             assert_eq!(f64::try_from(value1 % value2), Ok(2.0));
-    
+
             assert_eq!((value1 % 0.0).check(), Err(FloatError::NaN));
-    
+
             let value1 = UnguardedF64::new(6.0);
             assert_eq!((value1 % 0.0).check(), Err(FloatError::NaN));
-    
+
             let value1 = UnguardedF64::new(6.0);
             assert_eq!((value1 % f64::NAN).check(), Err(FloatError::NaN));
             assert_eq!((f64::NAN % value1).check(), Err(FloatError::NaN));
-    
+
             let value1 = UnguardedF64::new(6.0);
             assert_eq!((f64::INFINITY % value1).check(), Err(FloatError::Infinity));
             assert_eq!((value1 % f64::INFINITY).check(), Err(FloatError::Infinity));
-    
+
             let value1 = UnguardedF64::new(f64::INFINITY);
             let value2 = UnguardedF64::new(f64::NAN);
             assert_eq!((value1 % value2).check(), Err(FloatError::NaN));
